[step2] refactor: 결과 처리는 ResultProcessor에서 처리
