+++ src/main/kotlin/blackjack/domain/result/InitialDistributionResult.kt
@@ -4,4 +4,4 @@ import blackjack.domain.player.Players
 
 data class InitialDistributionResult(
     val players: Players
-)
+) : Result()
