+++ src/main/kotlin/blackjack/domain/result/Result.kt
@@ -1,0 +1,3 @@
+package blackjack.domain.result
+
+abstract class Result
